Modified scripts to choose project containing source images
Previously required source and target images to be in the same project. With this new change, the script is run in the project with the target images. Running the script will prompt the user to select the project with the source images. Assuming that the image naming convention is adhered to, the script will also automatically get the stain from the source images.

Also added the option to specify the separator in the image name. This change was done for added flexibility in case the image names did not use "_".